Create a Supervisor program for the pi
This commit is to add a supervisor program to the pi so that we can
program the axman can nodes to set each floor. by selecting option 4 you
will enter supervisor mode which will read the can bus for information
coming from the can nodes and send out the appropriate message to the
elevator controller to go to the correct floor. The code has been
cleaned up and the car was added so its buttons will now be registered.